Fix warnings

+++ src/tool/hpcrun/unwind/common/uw_hash.c
@@ -65,7 +65,7 @@
  * private operations
  ***************************************************************************/
 
-#define UW_HASH(key, size) (key < size ? key : (size_t)key % size);
+#define UW_HASH(key, size) (key < size ? key : key % size);
 
 /***************************************************************************
  * interface operations
@@ -87,7 +87,7 @@ void
 uw_hash_insert(uw_hash_table_t *uw_hash_table,
   void *key, ilmstat_btuwi_pair_t *ilm_btui, bitree_uwi_t *btuwi)
 {
-  size_t index = UW_HASH(key, uw_hash_table->size);
+  size_t index = UW_HASH(((size_t)key), uw_hash_table->size);
   uw_hash_entry_t *uw_hash_entry = &(uw_hash_table->uw_hash_entries[index]);
   uw_hash_entry->key = key;
   uw_hash_entry->ilm_btui = ilm_btui;
@@ -99,7 +99,7 @@ void
 uw_hash_ilm_btui_insert(uw_hash_table_t *uw_hash_table,
   void *key, ilmstat_btuwi_pair_t *ilm_btui)
 {
-  size_t index = UW_HASH(key, uw_hash_table->size);
+  size_t index = UW_HASH(((size_t)key), uw_hash_table->size);
   uw_hash_entry_t *uw_hash_entry = &(uw_hash_table->uw_hash_entries[index]);
   uw_hash_entry->ilm_btui = ilm_btui;
 }
@@ -109,7 +109,7 @@ void
 uw_hash_btuwi_insert(uw_hash_table_t *uw_hash_table,
   void *key, bitree_uwi_t *btuwi)
 {
-  size_t index = UW_HASH(key, uw_hash_table->size);
+  size_t index = UW_HASH(((size_t)key), uw_hash_table->size);
   uw_hash_entry_t *uw_hash_entry = &(uw_hash_table->uw_hash_entries[index]);
   uw_hash_entry->btuwi = btuwi;
 }
@@ -118,7 +118,7 @@ uw_hash_btuwi_insert(uw_hash_table_t *uw_hash_table,
 uw_hash_entry_t *
 uw_hash_lookup(uw_hash_table_t *uw_hash_table, void *key)
 {
-  size_t index = UW_HASH(key, uw_hash_table->size);
+  size_t index = UW_HASH(((size_t)key), uw_hash_table->size);
   uw_hash_entry_t *uw_hash_entry = &(uw_hash_table->uw_hash_entries[index]);
   if (uw_hash_entry->key != key) {
     return NULL;
@@ -143,7 +143,7 @@ uw_hash_delete_range(uw_hash_table_t *uw_hash_table, void *start, void *end)
 void
 uw_hash_delete(uw_hash_table_t *uw_hash_table, void *key)
 {
-  size_t index = UW_HASH(key, uw_hash_table->size);
+  size_t index = UW_HASH(((size_t)key), uw_hash_table->size);
   uw_hash_entry_t *uw_hash_entry = &(uw_hash_table->uw_hash_entries[index]);
   if (uw_hash_entry->key == key) {
     uw_hash_entry->key = NULL;
